Move old CRUD framework to the admin app
