+++ resources/assets/js/admin/components/forms/Quill.vue
